Refactor stringify() as TypeScript
