+++ src/stringify/stringify.ts
@@ -2,16 +2,32 @@ import { Alias } from '../ast/Alias.js'
 import { Node } from '../ast/Node.js'
 import { Pair } from '../ast/Pair.js'
 import { Scalar } from '../ast/Scalar.js'
+import type { Document } from '../doc/Document'
+import { Schema } from '../doc/Schema.js'
 import { stringifyString } from './stringifyString.js'
 
-function getTagObject(tags, item) {
+export interface StringifyContext {
+  anchors: Record<string, Node>
+  doc: Document
+  forceBlockIndent?: boolean
+  implicitKey?: boolean
+  indent: string
+  indentStep: string
+  indentAtStart?: number
+  inFlow?: boolean
+  stringify: typeof stringify
+  [key: string]: unknown
+}
+
+function getTagObject(tags: Schema.Tag[], item: Node) {
   if (item.tag) {
     const match = tags.filter(t => t.tag === item.tag)
     if (match.length > 0)
-      return match.find(t => t.format === item.format) || match[0]
+      return match.find(t => t.format === (item as Scalar).format) || match[0]
   }
 
-  let tagObj, obj
+  let tagObj: Schema.Tag | undefined = undefined
+  let obj: unknown
   if (item instanceof Scalar) {
     obj = item.value
     const match = tags.filter(t => t.identify && t.identify(obj))
@@ -23,6 +39,7 @@ function getTagObject(tags, item) {
   }
 
   if (!tagObj) {
+    // @ts-ignore
     const name = obj && obj.constructor ? obj.constructor.name : typeof obj
     throw new Error(`Tag not resolved for ${name} value`)
   }
@@ -30,7 +47,11 @@ function getTagObject(tags, item) {
 }
 
 // needs to be called before value stringifier to allow for circular anchor refs
-function stringifyProps(node, tagObj, { anchors, doc }) {
+function stringifyProps(
+  node: Node,
+  tagObj: Schema.Tag,
+  { anchors, doc }: StringifyContext
+) {
   const props = []
   const anchor = doc.anchors.getName(node)
   if (anchor) {
@@ -45,33 +66,38 @@ function stringifyProps(node, tagObj, { anchors, doc }) {
   return props.join(' ')
 }
 
-export function stringify(item, ctx, onComment, onChompKeep) {
-  const { schema } = ctx.doc
-
-  let tagObj
-  if (!(item instanceof Node)) {
-    item = ctx.doc.createNode(item, {
-      onTagObj: o => (tagObj = o),
-      wrapScalars: true
-    })
-  }
-
+export function stringify(
+  item: unknown,
+  ctx: StringifyContext,
+  onComment?: () => void,
+  onChompKeep?: () => void
+): string {
   if (item instanceof Pair) return item.toString(ctx, onComment, onChompKeep)
   if (item instanceof Alias) return item.toString(ctx)
-  if (!tagObj) tagObj = getTagObject(schema.tags, item)
 
-  const props = stringifyProps(item, tagObj, ctx)
+  let tagObj: Schema.Tag | undefined = undefined
+  const node: Node =
+    item instanceof Node
+      ? item
+      : ctx.doc.createNode(item, {
+          onTagObj: o => (tagObj = o),
+          wrapScalars: true
+        })
+
+  if (!tagObj) tagObj = getTagObject(ctx.doc.schema.tags, node)
+
+  const props = stringifyProps(node, tagObj, ctx)
   if (props.length > 0)
     ctx.indentAtStart = (ctx.indentAtStart || 0) + props.length + 1
 
   const str =
     typeof tagObj.stringify === 'function'
-      ? tagObj.stringify(item, ctx, onComment, onChompKeep)
-      : item instanceof Scalar
-      ? stringifyString(item, ctx, onComment, onChompKeep)
-      : item.toString(ctx, onComment, onChompKeep)
+      ? tagObj.stringify(node, ctx, onComment, onChompKeep)
+      : node instanceof Scalar
+      ? stringifyString(node, ctx, onComment, onChompKeep)
+      : node.toString(ctx, onComment, onChompKeep)
   if (!props) return str
-  return item instanceof Scalar || str[0] === '{' || str[0] === '['
+  return node instanceof Scalar || str[0] === '{' || str[0] === '['
     ? `${props} ${str}`
     : `${props}\n${ctx.indent}${str}`
 }
